Enqueue Styles
This plugin is compatible with the twenty seventeen Wordpress theme. It will change how posts are displayed on the front page of the twenty seventeen theme so that more of the posts are displayed at once and look more aesthetically pleasing. 

Problem Statement: Many websites powered by WordPress use a standard blog-style layout to display posts on the front page, with posts displayed in a list format. However, some users may prefer a more visually appealing layout, such as a grid layout, that better showcases their content. Therefore, there is a need for a WordPress plugin that can change the layout of posts on the front page of a site to a grid format, without requiring users to have advanced coding knowledge. 

Solution: Use enqueuestyles and custom CSS to change the format of the front page layout of a WordPress site. This approach involves creating a custom CSS file and enqueuing it using wp_enqueue_style() function in the plugin code. 

Directions for enduser: 

1. First download and ensure the Twenty Seventeen theme is applied. 
2. Download the Plugin. 
3. Activate the Plugin.
4. Generate Posts.
5. Posts will now be displayed in a Grid Format as cards for each post.